+++ src/measurements/DoubleExtensions.cs
@@ -1,4 +1,4 @@
-using Engineering.Units;
+using Engineering.Units.SI;
 
 namespace Engineering.Measurements
 {
