+++ src/units/expressions/IDerived{T}.cs
@@ -1,0 +1,7 @@
+namespace Engineering.Units.Expressions
+{
+    public interface IDerived<T> where T : IExpressible
+    {
+        Expression<T> Expression { get; }
+    }
+}
